+++ src/app/admin/create-user/page.jsx
@@ -1,8 +1,0 @@
-import CreateUserForm from "@/app/(components)/admin/CreateUserForm";
-import React from "react";
-
-const CreateUser = () => {
-  return <CreateUserForm />;
-};
-
-export default CreateUser;